% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cv_yaml.R
\name{read_cv}
\alias{read_cv}
\title{Read a yaml cv}
\usage{
read_cv(file, fileEncoding = "UTF-8", text, error.label, ...)
}
\arguments{
\item{file}{either a character string naming a file or a \link{connection}
    open for writing}

\item{fileEncoding}{character string: if non-empty declares the
    encoding used on a file (not a connection) so the character data can
    be re-encoded.  See \code{\link{file}}.}

\item{text}{character string: if \code{file} is not supplied and this is,
    then data are read from the value of \code{text} via a text connection.
    Notice that a literal string can be used to include (small) data sets
    within R code.}

\item{error.label}{a label to prepend to error messages (see Details).}

\item{...}{arguments to pass to \code{\link[yaml]{yaml.load}}}
}
\description{
Read a yaml cv
}
