% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_cv.R
\name{render_cv}
\alias{render_cv}
\title{Render a yaml cv to pdf}
\usage{
render_cv(
  x,
  output,
  overwrite = TRUE,
  template = system.file("templates", "awesomecv.Rnw", package = "yamlcv"),
  photo = NULL,
  photo_opts = c("rectangle", "edge", "right"),
  awesomecolor = FALSE,
  copy_tex = FALSE
)
}
\arguments{
\item{x}{\code{character} scalar. Path to a cv \code{.yaml} file}

\item{output}{\code{character} scalar. name of the output \code{.pdf} file}

\item{overwrite}{\code{logical} scalar. overwrite \code{output} if it exists?}

\item{template}{\code{character} scalar. Path to the template \code{.tex} file}

\item{photo}{\code{character} scalar. Path to a photo}

\item{photo_opts}{\code{character} vector. Options for placing the photo. Valid
values are \verb{circle|rectangle,edge/noedge,left/right}}

\item{awesomecolor}{color the first three letters of each section}

\item{copy_tex}{Copy the \code{.tex} file to the same output directory as
\code{output}}
}
\value{
\code{output}
}
\description{
Render a yaml cv to pdf
}
\section{Side effects}{

Creates the pdf file \code{output}
}

