% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cv_yaml.R
\name{render_skills}
\alias{render_skills}
\alias{render_person}
\alias{render_experience}
\alias{render_education}
\title{Render cv sections}
\usage{
render_skills(x)

render_person(x)

render_experience(x)

render_education(x)
}
\arguments{
\item{x}{a list with the appropriate elements (see \code{example-cv})}
}
\description{
Render cv sections
}
