% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yamlcv-package.R
\docType{package}
\name{yamlcv-package}
\alias{yamlcv}
\alias{yamlcv-package}
\title{yamlcv: Generate Beautiful CVs from 'yaml' Files}
\description{
Generate beautiful CVs from 'yaml' files. Currently
    only supports generating pdf CVs based on posquit0's Awesome-CV
    (https://github.com/posquit0/Awesome-CV) latex template, but in theory
    more templates and output formats could be supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s-fleck/yamlcv}
  \item Report bugs at \url{https://github.com/s-fleck/yamlcv/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
\keyword{internal}
